+++ member-service/src/main/java/org/example/memberservice/controller/MemberController.java
@@ -3,7 +3,7 @@ package org.example.memberservice.controller;
 import lombok.RequiredArgsConstructor;
 import org.example.common.apiPayload.response.ApiResponse;
 import org.example.common.constant.GatewayConstant;
-import org.example.memberservice.dto.MemberResponseDto;
+import org.example.memberservice.dto.member.MemberResponseDto;
 import org.example.memberservice.service.MemberService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
