--- member-service/src/main/java/org/example/memberservice/dto/OcrRequestDto.java
@@ -1,4 +1,0 @@
-package org.example.memberservice.dto;
-
-public class OcrRequestDto {
-}
