+++ member-service/src/main/java/org/example/memberservice/dto/ocr/OcrRequestDto.java
@@ -1,0 +1,4 @@
+package org.example.memberservice.dto.ocr;
+
+public class OcrRequestDto {
+}
