+++ member-service/src/main/java/org/example/memberservice/dto/ocr/OcrResponseDto.java
@@ -1,0 +1,4 @@
+package org.example.memberservice.dto.ocr;
+
+public class OcrResponseDto {
+}
