+++ member-service/src/main/java/org/example/memberservice/dto/publicProfile/PublicProfileResponseDto.java
@@ -1,4 +1,4 @@
-package org.example.memberservice.dto;
+package org.example.memberservice.dto.publicProfile;
 
 import org.example.memberservice.domain.Member;
 import org.example.memberservice.domain.PublicProfile;
