+++ member-service/src/main/java/org/example/memberservice/service/MemberPreferenceService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.memberservice.domain.Member;
 import org.example.memberservice.domain.MemberPreference;
-import org.example.memberservice.dto.MemberPreferencesRequestDto;
-import org.example.memberservice.dto.MemberPreferencesResponseDto;
+import org.example.memberservice.dto.memberPreferences.MemberPreferencesRequestDto;
+import org.example.memberservice.dto.memberPreferences.MemberPreferencesResponseDto;
 import org.example.memberservice.repository.MemberPreferenceRepository;
 import org.example.memberservice.repository.MemberRepository;
 import org.springframework.stereotype.Service;
