+++ member-service/src/main/java/org/example/memberservice/service/OcrService.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.memberservice.domain.Member;
-import org.example.memberservice.dto.MemberInfoDto;
+import org.example.memberservice.dto.member.MemberInfoDto;
 import org.example.memberservice.repository.MemberRepository;
 import org.example.memberservice.util.CryptoUtil;
 import org.springframework.beans.factory.annotation.Value;
