+++ member-service/src/main/java/org/example/memberservice/service/PublicProfileService.java
@@ -4,9 +4,9 @@ package org.example.memberservice.service;
 import lombok.RequiredArgsConstructor;
 import org.example.memberservice.domain.Member;
 import org.example.memberservice.domain.PublicProfile;
-import org.example.memberservice.dto.PublicProfileRequestDto;
-import org.example.memberservice.dto.PublicProfileResponseDto;
-import org.example.memberservice.dto.PublicProfileUpdateRequestDto;
+import org.example.memberservice.dto.publicProfile.PublicProfileRequestDto;
+import org.example.memberservice.dto.publicProfile.PublicProfileResponseDto;
+import org.example.memberservice.dto.publicProfile.PublicProfileUpdateRequestDto;
 import org.example.memberservice.repository.MemberRepository;
 import org.example.memberservice.repository.PublicProfileRepository;
 import org.springframework.stereotype.Service;
